python-telegram-bot/JokeBot/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.2.25
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
multidict==6.7.1
propcache==0.4.1
python-dotenv==1.2.2
python-telegram-bot==22.6
typing_extensions==4.15.0
yarl==1.23.0
